Clean up zWorld and zWorld2 + match most of the rest of them

+++ src/Speed/Indep/Src/AI/Common/AIPursuit.cpp
@@ -551,7 +551,7 @@ void AIPursuit::OnAttached(IAttachable *pOther) {
 
             IPerpetrator *iperp;
             if (mTarget->QueryInterface(&iperp) && mRepPointsPerMinute == 0) {
-                int perpHeat = iperp->GetHeat();
+                int perpHeat = static_cast<int>(iperp->GetHeat());
                 if (plevels) {
                     mRepPointsPerMinute = plevels->RepPointsPerMinute();
                 }
